requirements.txt
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.8.0
idna==3.10
pymongo==4.15.2
python-dotenv==1.1.1
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
